The same C&V S&R job everyone has to do to get a ParseInt([]byte) function
since the standard library's strconv doesn't have that direction, and in
go 1.7 converting a []byte to a string to pass into strconv.ParseInt()
creates a copy of the []byte for the string, which then has to be
garbage collected.

The suggestion from on high is to wait for the compiler to be able to
optimize away the copy of []byte to string. However I need fast code
today.